Fix BoundedQueueExecutor and StreamingDataflowWorker to actually limit memory from windmill
Currently, because the queue is only limited by number of elements, there can be up to (num threads + queue size) elements outstanding at a time, which for large work items will almost certainly OOM the worker.

This change both makes this limit explicit and adds a 500 MB memory limit on outstanding WorkItems to push back on windmill before workers run out of memory.

+++ runners/google-cloud-dataflow-java/src/main/java/org/apache/beam/runners/dataflow/options/DataflowPipelineDebugOptions.java
@@ -177,10 +177,14 @@ public interface DataflowPipelineDebugOptions extends ExperimentalOptions, Pipel
 
   void setNumberOfWorkerHarnessThreads(int value);
 
-  /** Maximum number of bundles outstanding from windmill before the worker stops requesting. */
+  /**
+   * Maximum number of bundles outstanding from windmill before the worker stops requesting.
+   *
+   * <p>If <= 0, use the default value of 100 + getNumberOfWorkerHarnessThreads()
+   */
   @Description(
       "Maximum number of bundles outstanding from windmill before the worker stops requesting.")
-  @Default.Integer(400)
+  @Default.Integer(0)
   int getMaxBundlesFromWindmillOutstanding();
 
   void setMaxBundlesFromWindmillOutstanding(int value);
